requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
pytest==8.3.5
httpx==0.28.1
